Install libncurses5 and libtinfo5 for ghc < 8.3 on Linux (#136)
Fixes #135.  
Adds a new package dependency `compare-versions` for robustly checking whether the GHC we seek to install is below 8.3.
Ideally, we would only try to install `libncurses5` and `libtinfo5` on `ubuntu-22.04`, but I didn't see a way to check whether we are exactly on this virtual environment, so we try it always under `Linux` for ghc < 8.3.  Seems like in the worst case, this fails and does nothing (fingers crossed).

Also in this commit: fixing the check whether `stack` is above a given minimum version: now also uses `compareVersions`.